python/requirements.txt
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
idna==3.7
isodate==0.6.1
lxml==5.2.2
numpy==1.26.4
onvif_zeep==0.2.12
opencv-python==4.10.0.82
platformdirs==4.2.2
pytz==2024.1
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
six==1.16.0
urllib3==2.2.1
zeep==4.2.1
